requirements.txt
python-telegram-bot==20.8
python-dotenv==1.0.0
SQLAlchemy==2.0.27
asyncmy==0.2.8
greenlet==3.1.1